apps/tts-service/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# TTS Engine
piper-tts==1.2.0
onnxruntime-gpu==1.16.3

# Audio Processing
soundfile==0.12.1
numpy>=1.24.0
pydub==0.25.1

# Async & Websockets
websockets==12.0
python-multipart==0.0.6

# Caching (optional)
redis==5.0.1

# Utilities
python-dotenv==1.0.0
